Windows 10 x86_64, mingw-w64-6.0.0, gcc-8.3.0-i686, Qt-5.9.7-i686
winlib: update script, add source and binary files

+++ winlibs/include/ogg/config_types.h
@@ -16,10 +16,10 @@
 #  include <sys/types.h>
 #endif
 
-typedef short ogg_int16_t;
-typedef unsigned short ogg_uint16_t;
-typedef int ogg_int32_t;
-typedef unsigned int ogg_uint32_t;
-typedef long long ogg_int64_t;
+typedef int16_t ogg_int16_t;
+typedef uint16_t ogg_uint16_t;
+typedef int32_t ogg_int32_t;
+typedef uint32_t ogg_uint32_t;
+typedef int64_t ogg_int64_t;
 
 #endif
